+++ geotrek/signage/sql/00_cleanup.sql
@@ -1,0 +1,2 @@
+-- Cleanup public schema if installed before release 0.28
+DROP VIEW IF EXISTS public.a_v_signaletique;