+++ geotrek/tourism/tests/test_parsers.py
@@ -75,7 +75,7 @@ class ParserTests(TranslationResetMixin, TestCase):
         mocked.return_value.json = mocked_json
         FileType.objects.create(type=u"Photographie")
         self.assertEqual(TouristicEvent.objects.count(), 0)
-        call_command('import', 'geotrek.tourism.parsers.TouristicEventApidaeParser', verbosity=0)
+        call_command('import', 'geotrek.tourism.parsers.TouristicEventApidaeParser', verbosity=2)
         self.assertEqual(TouristicEvent.objects.count(), 1)
         event = TouristicEvent.objects.get()
         self.assertEqual(event.eid, u"323154")